Create outgoing activity in outer scope of the pipeline (IOutgoingLogicalMessageContext) to allow customization. Resolves jbogard/NServiceBus.Extensions.Diagnostics#9
on-behalf-of: @Phreesia <mitchell.etter@phreesia.com
